connect db scheme to api
