+++ modules/fbs-core/web/src/app/page-components/sql-playground/db-scheme/db-scheme-routines/db-scheme-routines.component.ts
@@ -1,7 +1,6 @@
 import { Component, Input, OnInit } from "@angular/core";
 import { Routine } from "src/app/model/sql_playground/Routine";
 import { DbSchemeComponent } from "../db-scheme.component";
-import routinesJson from "../test-data/routines.json";
 
 @Component({
   selector: "app-db-scheme-routines",
@@ -12,9 +11,5 @@ export class DbSchemeRoutinesComponent
   extends DbSchemeComponent
   implements OnInit
 {
-  @Input() routines: Routine[] = routinesJson.SQLRoutines;
-
-  ngOnInit(): void {
-    console.log("");
-  }
+  @Input() routines: Routine[];
 }
