+++ modules/fbs-core/web/src/app/page-components/sql-playground/db-scheme/db-scheme-views/db-scheme-views.component.ts
@@ -1,7 +1,6 @@
 import { Component, Input, OnInit } from "@angular/core";
 import { View } from "src/app/model/sql_playground/View";
 import { DbSchemeComponent } from "../db-scheme.component";
-import viewsJson from "../test-data/views.json";
 
 @Component({
   selector: "app-db-scheme-views",
@@ -12,9 +11,5 @@ export class DbSchemeViewsComponent
   extends DbSchemeComponent
   implements OnInit
 {
-  @Input() views: View[] = viewsJson.SQLViews;
-
-  ngOnInit(): void {
-    console.log("");
-  }
+  @Input() views: View[];
 }
