+++ modules/fbs-core/web/src/app/page-components/sql-playground/db-scheme/db-scheme.component.ts
@@ -10,7 +10,5 @@ export class DbSchemeComponent implements OnInit {
 
   constructor() {}
 
-  ngOnInit(): void {
-    console.log("");
-  }
+  ngOnInit(): void {}
 }
