+++ modules/fbs-core/web/src/app/page-components/sql-playground/sql-playground.component.html
@@ -26,12 +26,22 @@
   </div>
   <div class="db-schema-grid">
     <app-bordered-container [yScrollable]="true" [title]="'DB Schema'">
-      <app-db-scheme-table [title]="'Tabellen'"></app-db-scheme-table>
-      <app-db-scheme-views [title]="'Views'"></app-db-scheme-views>
+      <app-db-scheme-table
+        [title]="'Tabellen'"
+        [tables]="tables"
+      ></app-db-scheme-table>
+      <app-db-scheme-views
+        [title]="'Views'"
+        [views]="views"
+      ></app-db-scheme-views>
       <app-db-scheme-routines
         [title]="'Stored Procedures'"
+        [routines]="routines"
       ></app-db-scheme-routines>
-      <app-db-scheme-triggers [title]="'Triggers'"></app-db-scheme-triggers>
+      <app-db-scheme-triggers
+        [title]="'Triggers'"
+        [triggers]="triggers"
+      ></app-db-scheme-triggers>
     </app-bordered-container>
   </div>
 </div>
